requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
autopep8==2.3.1
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
clicksend-client==5.0.78
Django==5.1.1
django-filter==24.3
django-otp==1.5.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
frozenlist==1.4.1
idna==3.10
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
twilio==9.3.3
urllib3==2.2.3
yarl==1.13.1
